SimpleChatTC:SimpleProxy:Move pdf logic into its own module
